Clean up table file structure
