+++ webview/src/experiments/components/Experiments.tsx
@@ -27,7 +27,7 @@ import styles from './table/styles.module.scss'
 import { AddColumns, Welcome } from './GetStarted'
 import { RowSelectionProvider } from './table/RowSelectionContext'
 import { CellValue } from './table/content/Cell'
-import { CellSecondaryName } from './table/CellSecondaryName'
+import { CellSecondaryName } from './table/body/CellSecondaryName'
 import { AddStage } from './AddStage'
 import { buildColumns, columnHelper } from '../util/buildColumns'
 import { sendMessage } from '../../shared/vscode'
