+++ webview/src/experiments/components/table/RowSelectionContext.tsx
@@ -1,5 +1,5 @@
 import React, { createContext, useState } from 'react'
-import { RowProp } from './interfaces'
+import { RowProp } from '../../util/interfaces'
 
 export interface RowSelectionContextValue {
   selectedRows: Record<string, RowProp | undefined>
