+++ webview/src/experiments/components/table/body/Cell.tsx
@@ -1,12 +1,12 @@
 import { flexRender } from '@tanstack/react-table'
 import React, { ReactNode } from 'react'
 import cx from 'classnames'
-import styles from './styles.module.scss'
-import { CellProp, RowProp } from './interfaces'
 import { CellRowActionsProps, CellRowActions } from './CellRowActions'
-import { CellValue, isValueWithChanges } from './content/Cell'
-import { clickAndEnterProps } from '../../../util/props'
-import { ErrorTooltip } from '../../../shared/components/tooltip/ErrorTooltip'
+import styles from '../styles.module.scss'
+import { CellValue, isValueWithChanges } from '../content/Cell'
+import { CellProp, RowProp } from '../../../util/interfaces'
+import { clickAndEnterProps } from '../../../../util/props'
+import { ErrorTooltip } from '../../../../shared/components/tooltip/ErrorTooltip'
 
 const cellHasChanges = (cellValue: CellValue) =>
   isValueWithChanges(cellValue) ? cellValue.changes : false
