+++ webview/src/experiments/components/table/body/NestedRow.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
-import styles from './styles.module.scss'
 import { BatchSelectionProp, RowContent } from './Row'
-import { RowProp } from './interfaces'
+import styles from '../styles.module.scss'
+import { RowProp } from '../../../util/interfaces'
 
 export const NestedRow: React.FC<RowProp & BatchSelectionProp> = ({
   row,
