+++ webview/src/experiments/components/table/body/Row.tsx
@@ -9,16 +9,16 @@ import {
 } from 'dvc/src/experiments/webview/contract'
 import { MessageFromWebviewType } from 'dvc/src/webview/contract'
 import { EXPERIMENT_WORKSPACE_ID } from 'dvc/src/cli/dvc/contract'
-import { RowProp } from './interfaces'
-import styles from './styles.module.scss'
 import { FirstCell, CellWrapper } from './Cell'
-import { RowSelectionContext } from './RowSelectionContext'
 import { RowContextMenu } from './RowContextMenu'
-import { sendMessage } from '../../../shared/vscode'
-import { ContextMenu } from '../../../shared/components/contextMenu/ContextMenu'
-import { HandlerFunc } from '../../../util/props'
-import { cond } from '../../../util/helpers'
-import { ExperimentsState } from '../../store'
+import { RowProp } from '../../../util/interfaces'
+import styles from '../styles.module.scss'
+import { RowSelectionContext } from '../RowSelectionContext'
+import { sendMessage } from '../../../../shared/vscode'
+import { ContextMenu } from '../../../../shared/components/contextMenu/ContextMenu'
+import { HandlerFunc } from '../../../../util/props'
+import { cond } from '../../../../util/helpers'
+import { ExperimentsState } from '../../../store'
 
 const getExperimentTypeClass = ({ status, selected }: Experiment) => {
   if (isRunning(status)) {
